move PruneLiteralsInOrderBy out of core
